[CAL][9] - Implement CommitReportsGTETimestamp (#1000)
