+++ pkg/chainaccessor/types.go
@@ -1,10 +1,53 @@
 package chainaccessor
 
-import cciptypes "github.com/smartcontractkit/chainlink-ccip/pkg/types/ccipocr3"
+import (
+	"math/big"
 
-// SendRequestedEvent represents the contents of the event emitted by the CCIP onramp when a message is sent.
+	cciptypes "github.com/smartcontractkit/chainlink-ccip/pkg/types/ccipocr3"
+)
+
+// ---------------------------------------------------
+// The following types match the structs defined in the EVM contracts are used to decode these
+// on-chain events. They should eventually be replaced by chain-reader modifiers and use the
+// base cciptypes.CommitReport type.
+
+// SendRequestedEvent represents the contents of the event emitted by the CCIP OnRamp when a
+// message is sent.
 type SendRequestedEvent struct {
 	DestChainSelector cciptypes.ChainSelector
 	SequenceNumber    cciptypes.SeqNum
 	Message           cciptypes.Message
+}
+
+// CommitReportAcceptedEvent represents the contents of the event emitted by the CCIP OffRamp when a
+// commit report is accepted.
+type CommitReportAcceptedEvent struct {
+	BlessedMerkleRoots   []MerkleRoot
+	UnblessedMerkleRoots []MerkleRoot
+	PriceUpdates         PriceUpdates
+}
+
+type MerkleRoot struct {
+	SourceChainSelector uint64
+	OnRampAddress       cciptypes.UnknownAddress
+	MinSeqNr            uint64
+	MaxSeqNr            uint64
+	MerkleRoot          cciptypes.Bytes32
+}
+
+type TokenPriceUpdate struct {
+	SourceToken cciptypes.UnknownAddress
+	UsdPerToken *big.Int
+}
+
+type GasPriceUpdate struct {
+	// DestChainSelector is the chain that the gas price is for (some plugin source chain).
+	// Not the chain that the gas price is stored on.
+	DestChainSelector uint64
+	UsdPerUnitGas     *big.Int
+}
+
+type PriceUpdates struct {
+	TokenPriceUpdates []TokenPriceUpdate
+	GasPriceUpdates   []GasPriceUpdate
 }
